refactor: 课程平均评分改为使用Integer存储

+++ learning-service/learning-course/src/main/java/cn/linter/learning/course/dao/EvaluationDao.java
@@ -67,6 +67,6 @@ public interface EvaluationDao {
      * @param courseId 课程ID
      * @return 平均评分
      */
-    Short selectAverageScoreByCourseId(Long courseId);
+    Integer selectAverageScoreByCourseId(Long courseId);
 
 }
